pyproject.toml
[project]
name = "argparser-examples"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.3",
]
